requirements.txt
# Core
python-dotenv>=1.0.0

# LLM SDK
google-generativeai>=0.3.0
openai>=1.0.0

# Document Processing
python-docx>=1.1.0

# UI
gradio>=4.0.0
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Vector Database
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Text-to-Speech
gTTS>=2.5.0
pydub>=0.25.1

# HuggingFace Transformers (STT with Whisper)
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0
accelerate>=0.24.0
librosa>=0.10.0
faster-whisper>=1.0.0

# Speaker Diarization
pyannote.audio>=3.1.0
speechbrain>=0.5.16

# LangChain & Advanced RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-pinecone>=0.1.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
pinecone-client>=3.0.0
tiktoken>=0.5.0

# Vosk (Offline STT - Free & Lightweight - Backup option)
vosk>=0.3.45
soundfile>=0.12.1

# Data Processing
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
